[kmh22] feat: reformatted code for merge, re-made page.tsx
