+++ web-app/src/app/page.tsx
@@ -1,123 +1,101 @@
-"use client";
-
-import { useEffect } from "react";
-import Script from "next/script";
 import Image from "next/image";
-import WebGazerClient from "./WebGazerClient"; // We'll keep a separate file
-import Calibration from "./Calibration";
 
 export default function Home() {
-  useEffect(() => {
-    // By now, the script has loaded with "beforeInteractive".
-    const webgazer = (window as any).webgazer;
-    if (!webgazer) {
-      console.error("WebGazer is not defined on window.");
-      return;
-    }
-    console.log("WebGazer found. Setting up...");
-
-    // Example gaze callback
-    function updateHighlight(data: { x: number; y: number } | null) {
-      if (!data) return;
-      console.log("Gaze data:", data);
-
-      // Quarter-highlighting logic
-      const screenWidth = window.innerWidth;
-      const quarterWidth = screenWidth / 4;
-      let activeQuarter = 0;
-      if (data.x < quarterWidth) activeQuarter = 1;
-      else if (data.x < quarterWidth * 2) activeQuarter = 2;
-      else if (data.x < quarterWidth * 3) activeQuarter = 3;
-      else activeQuarter = 4;
-
-      for (let i = 1; i <= 4; i++) {
-        const quarter = document.getElementById(`quarter-${i}`);
-        if (quarter) {
-          quarter.style.opacity = i === activeQuarter ? "0.5" : "0";
-        }
-      }
-
-      // Move our custom highlight dot
-      const highlight = document.getElementById("highlight");
-      if (highlight) {
-        highlight.style.left = `${data.x}px`;
-        highlight.style.top  = `${data.y}px`;
-      }
-    }
-
-    // Begin the WebGazer pipeline & set the listener
-    webgazer
-      .setGazeListener(updateHighlight)
-      .begin()
-      .then(() => console.log("WebGazer has started successfully!"))
-      .catch((err: any) => console.error("WebGazer failed to start:", err));
-
-    // Cleanup on unmount
-    return () => {
-      webgazer.clearGazeListener();
-      console.log("WebGazer listener cleared.");
-    };
-  }, []);
-
   return (
-    <>
-      {/* Load WebGazer before everything else */}
-      <Script
-        src="https://webgazer.cs.brown.edu/webgazer.js"
-        strategy="beforeInteractive" // ensures script is loaded early
-        onLoad={() => console.log("WebGazer script loaded (beforeInteractive)!")}
-        onError={() => console.error("Failed to load WebGazer script.")}
-      />
-
-      {/* Main UI */}
-      <div className="relative min-h-screen">
-        {/* Our own highlight dot (second red dot, distinct from the built-in one) */}
-        <div
-          id="highlight"
-          style={{
-            position: "absolute",
-            width: "20px",
-            height: "20px",
-            background: "red",
-            borderRadius: "50%",
-            pointerEvents: "none",
-            top: 0,
-            left: 0,
-            transform: "translate(-50%, -50%)",
-            transition: "top 0.1s ease, left 0.1s ease",
-            zIndex: 1000,
-          }}
+    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
+      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
+        <Image
+          className="dark:invert"
+          src="/next.svg"
+          alt="Next.js logo"
+          width={180}
+          height={38}
+          priority
         />
+        <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
+          <li className="mb-2">
+            Hello everyone, this is
+            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-semibold">
+              src/app/page.tsx
+            </code>
+            .
+          </li>
+          <li>Save and see your changes instantly.</li>
+        </ol>
 
-        {/* The four quarters */}
-        <div
-          id="quarter-1"
-          className="absolute top-0 left-0 h-full w-1/4 bg-green-500 opacity-0 pointer-events-none"
-        />
-        <div
-          id="quarter-2"
-          className="absolute top-0 left-1/4 h-full w-1/4 bg-green-500 opacity-0 pointer-events-none"
-        />
-        <div
-          id="quarter-3"
-          className="absolute top-0 left-1/2 h-full w-1/4 bg-green-500 opacity-0 pointer-events-none"
-        />
-        <div
-          id="quarter-4"
-          className="absolute top-0 left-3/4 h-full w-1/4 bg-green-500 opacity-0 pointer-events-none"
-        />
-
-        {/* Our separate client file where we can configure WebGazer params */}
-        <WebGazerClient />
-
-        <main>
-          <Image src="/next.svg" alt="Next.js logo" width={180} height={38} />
-          <p>Hi from page.tsx</p>
-        </main>
-
-        {/* Optional calibration */}
-        <Calibration />
-      </div>
-    </>
+        <div className="flex gap-4 items-center flex-col sm:flex-row">
+          <a
+            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
+            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Image
+              className="dark:invert"
+              src="/vercel.svg"
+              alt="Vercel logomark"
+              width={20}
+              height={20}
+            />
+            Deploy now
+          </a>
+          <a
+            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
+            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Read our docs
+          </a>
+        </div>
+      </main>
+      <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/file.svg"
+            alt="File icon"
+            width={16}
+            height={16}
+          />
+          Learn
+        </a>
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/window.svg"
+            alt="Window icon"
+            width={16}
+            height={16}
+          />
+          Examples
+        </a>
+        <a
+          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
+          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Image
+            aria-hidden
+            src="/globe.svg"
+            alt="Globe icon"
+            width={16}
+            height={16}
+          />
+          Go to nextjs.org →
+        </a>
+      </footer>
+    </div>
   );
 }
