+++ web-app/src/app/webGazer/WebGazerClient.tsx
